+++ airlineroute/foundation/src/main/java/com/asid/foundation/datastructure/graph/CustomDirectedWeightGraphAdapter.java
@@ -195,7 +195,7 @@ public class CustomDirectedWeightGraphAdapter <V, E extends DefaultEdge<V>> exte
 
     @Override
     public void setEdgeWeight(E e, double v) {
-        e.setWeight((long) v);
+        e.setWeight(v);
     }
 
     @Override
@@ -236,7 +236,7 @@ public class CustomDirectedWeightGraphAdapter <V, E extends DefaultEdge<V>> exte
         return outgoingEdges;
     }
 
-    public E addEdge(V v, V v1, long weight) {
+    public E addEdge(V v, V v1, double weight) {
         if(containsVertex(v) && containsVertex(v1)){
             E edge = edgeFactory.createEdge(v, v1);
             edge.setWeight(weight);
